Fix test checking notifications without a link
The test was supposed to be testing notifications without a link, but it
was adding a link. It was only passing because the expectations was
checking whether there was a link whose text was a URL, but what we
want to check here is whether the URL is present in the `href`
attribute.

We're about to delete the page showing public stats and, since this link
was using the `/stats` path, we're fixing it.