Copy original files to custom folder
Note that we are overwriting the entire VotationType model and
this is not the usual way to customise Consul. The reason is
that in the next commit we want to add a new value in the enum
:vote_type and until we upgrade to rails 7.1 we won't be able
to do this without having to copy the whole file.

With the rails version 7.1 we can add in the
`app/models/custom/votation_type.rb` the following code:

```
enum :vote_type, %w[unique multiple open], instance_methods: false
```