Remove redundant files in docs subfolder
These files were both in the Consul Democracy repository and the
documentation repository. Since now both repositories have been merged
into one, we can remove the duplicate files.

Note there was a small difference in the `.mdlrc` file because we
changed it in commit 4656af401. We're keeping this change so we don't
get any syntax warnings in the documentation.